add debug msgs and allow manual preferences for playing around

+++ main.cpp
@@ -7,9 +7,10 @@ using namespace std;
 // test
 int main(int argc, char** argv)
 {
-  int prefType = 1;
+  float prefType = 1;
   if (argc >= 2) {
-    prefType = atoi(argv[1]);
+    prefType = atof(argv[1]);
+
   }
   vector<int> examplar = affinityPropagation(stdin, prefType);
   for (size_t i = 0; i < examplar.size(); ++i) {
